feat: v4 dark theme
