+++ components/theme/export.ts
@@ -2,7 +2,7 @@
 import { useToken as useInternalToken } from '.';
 import defaultAlgorithm from './themes/default';
 import darkAlgorithm from './themes/dark';
-import v4Algorithm from './themes/v4';
+import { defaultAlgorithmV4, darkAlgorithmV4 } from './themes/v4';
 
 // ZombieJ: We export as object to user but array in internal.
 // This is used to minimize the bundle size for antd package but safe to refactor as object also.
@@ -18,5 +18,6 @@ export default {
   useToken,
   defaultAlgorithm,
   darkAlgorithm,
-  v4Algorithm,
+  defaultAlgorithmV4,
+  darkAlgorithmV4,
 };
