+++ site/theme/template/Layout/index.tsx
@@ -253,8 +253,10 @@ export default class Layout extends React.Component<LayoutPropsType, LayoutState
     switch (this.state.v5theme) {
       case 'dark':
         return antdTheme.darkAlgorithm;
-      case 'v4':
-        return antdTheme.v4Algorithm;
+      case 'defaultV4':
+        return antdTheme.defaultAlgorithmV4;
+      case 'darkV4':
+        return antdTheme.darkAlgorithmV4;
       case 'default':
       default:
         return antdTheme.defaultAlgorithm;
